Fix handling of @external on interface fields
This commit:
- rejects @external on interface fields in fed2 schema.
- removes @external on inteface fields in fed1 schema when
  auto-migrating.

The rational is that for fed2, @external is about fields that are not
resolved by the subgraph, but interface fields are not resolved in the
first place so @external makes no sense on them and "should" be rejected
(or rather, accepting them would just muddy the water; note that they
were mostly rejected already due to rejecting unused @external and not
allowing federation directives on interface, but the new message is a
clearer). But fed1 required some interface fields to be marked @external
(without that impacting any runtime behavior), so we make sure migration
get rid of it.